+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changes since 0.6.0 onwards.
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+0.9.7 (2020-12-24)
+~~~~~~~~~~~~~~~~~~
+* Added metadata to dataset (#93)
+* Added correct_ramp to dataset
+* Added ramp parameter to dataset.fit_transit and fit_eclipse.
+* Added ramp to multivisit
+* AIC and BIC now correct and consistent for all fitting methods
+* Improve selction of min/max values for parameters in Multivisit
+* Changes for DRP13 compatibility.
+
 0.9.6 (2020-11-24)
 ~~~~~~~~~~~~~~~~~~
 * Force user to set decontaminate option in dataset.get_lightcurve()